+ field specifier is fixed string unless -r/--regex
add a new option to treat the field pattern as a regex rather than a
fixed string.  then make the default behavior if unspecified to use
space or tab as a [regex] pattern.  however, if specified, the default
will be to treat it as a fixed string, so that giving e.g. '.' as the
indelim will have expected behavior and not just match any character,
which has unexpected results.  but not supplying it will keep the
expected behavior of any whitespace being a field separator

this fixes bug 0d7a57a0