/ hexversion: move version check early in script to see error
python interprets lines while it reads them.  in our code, we use things
like '/' as function parameters, which are invalid before python 3.8,
for example.  so if the version check is in __main__ block, which is at
the *end* of the file, we will get syntax error parsing the file before
we got a chance to do a version check.

so we'll move version check very early, as the first code, even before
imports (which maybe are different in later python versions).  of
course, we still need to import hexversion and exit...

this does also mean we have to do error print and exit ourselves because
bomb() has not yet been defined