+ irs, ors, ifs, ofs and regex forms can now be specified as opts
allow the full set to be expressed as options.  previously we could not
even specify record separators.  using the 1,2,3 form of positionally
specified arguments it was not possible without adding more positions,
which would have been unwieldy and hard to remember, and require
specifying all params even if we only wanted to give just one (if the
param was at the end of the list)

we still keep the positionally-specified way possible as a shortcut,
since it's arguably the common case for usage and will help with cli
speed, but that can only be used if ifs and ofs are the only ones to
specify

in the future when we take files, this form of specification has
uncertain future.  we can either say "if you want to specify files, you
must use the option form, and then all args after the first (which is
rangelist) are filenames", or allow specification of ifs/ofs
positionally but require a '--' if filenames will be given after it.  or
maybe both: if options are specified at all (besides -r), positional is
disabled, and all args assumed to be filenames, but if no options, then
they are considered to be ifs/ofs and filenames cannot be given, unless
one of them is '--' before the third, in which case one and two are
taken as ifs/ofs