- discard empty last record, ie made by trailing NL in text files
all files typically end in a trailing delimiter, but don't really mean
it forms the edge of a final empty record.  so we discard that.  this
mirrors awk.

it's interesting that this same logic isn't used on fields, ie a
trailing delimiter on a line does create an extra empty field in awk.
but we will not question that logic here, rather go by expectation of
tools written decades ago and whose logic has already been reasoned
about by many others.