+ field: use buffered stdout writes, rather than os.write()
originally we had switched to os.write() because sys.stdout is actually
a TextIOBase but if we access its .buffer method is an
_io.BufferedWriter which take bytes and manages buffering efficiently.
we have no particular need to use the lower-level os.write(), and
buffered output should be more efficient anyways.  we can do flushes
ourselves if we want to do record-buffered output, based on a cli option
(forthcoming).  conversely, when we use os.write(), it would always
write record-sized buffers, unless we wrote our own buffering...