/ field: noblanks: fix infinite loop if a blank record was actually hit
we gave no chance to exit the loop if --noblanks and one was found, we
should have kept going, but skipped the action, rather than just looping
endlessly

yes, we should have a test suite, which would have caught this easily.
guessing it never worked in the first place, it was just an obvious
thing to implement so we put it in.

ideally we wouldn't have these conditional branches and the special
first iteration of the loop, but rather move the record-read to the
beginning of the loop and just 'continue' if we're to skip.  this would
make for a lot simpler code.  however, after trying for some time to
implement it, was not able to do it that way.  the problem is, there's
stuff done both before and after the record read, so we would have had
just as many conditions but inverted everything.  at least this way the
logic is more clear, if annoying that we have to test conditions every
loop.  so it goes...