/ dmain: finally: at the very end dup2 stderr to /dev/null
this seems to be required but it's not clear why, there's some
speculation in the comment in the commit.  it's a bit silly that we have
to go through these theatrics but it seems to be the case.  maybe we
should try having a signal handler and using os._exit and doing things
ourselves